fallback to rust codeblock if in rust file
don't know if this can be bad, but for some reasons rust_analyzer don't
return rust as a language to the codeblock for signature help, while it
does for hover information. add rust as fallback to make this similar
for now. might make it smarter or revert later if this is bad